requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
numpy==1.14.5
PyMySQL==0.9.2
python-dateutil==2.7.3
requests==2.19.1
SQLAlchemy==1.2.11
